+++ packages/react/popper/src/index.ts
@@ -7,3 +7,5 @@ export { PopperItem } from "./popper-item";
 export { PopperLabel } from "./popper-label";
 export { PopperSeparator } from "./popper-separator";
 export { PopperTrigger } from "./popper-trigger";
+
+export type * from "./popper.types";