+++ packages/react/popper/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @everest-ui/react-popper
 
+## 1.4.0
+
+### Minor Changes
+
+- Remove unnecessary packages
+
 ## 1.3.1
 
 ### Patch Changes